[MWPW-158382] [Performance] Merging scripts and utils and removing styles css (#433)
* moving scripts to head

* moving scripts to head

* added preloads

* removed styles and preloads

* changes

* changes

* eslint

* modifed as per review changes

---------
