+++ 404.html
@@ -8,7 +8,10 @@
   <meta name="header" content="global-navigation" />
   <meta name="footer" content="global-footer" />
   <meta name="viewport" content="width=device-width, initial-scale=1" />
-  <script src="/creativecloud/scripts/scripts.js" type="module"></script>
+  <script type="module">
+    import { scriptInit } from '/creativecloud/scripts/utils.js';
+    scriptInit();
+  </script>
   <link rel="icon" href="data:," />
 </head>
