+++ creativecloud/scripts/utils.js
@@ -16,6 +16,145 @@
  */
 const COOKIE_SIGNED_IN = 'acomsis';
 const COOKIE_SIGNED_IN_STAGE = 'acomsis_stage';
+const CHINA_SIGNED_IN_HOME_PATH = '/cn/creativecloud/roc/home';
+
+const locales = {
+  // Americas
+  ar: { ietf: 'es-AR', tk: 'oln4yqj.css' },
+  br: { ietf: 'pt-BR', tk: 'inq1xob.css' },
+  ca: { ietf: 'en-CA', tk: 'pps7abe.css' },
+  ca_fr: { ietf: 'fr-CA', tk: 'vrk5vyv.css' },
+  cl: { ietf: 'es-CL', tk: 'oln4yqj.css' },
+  co: { ietf: 'es-CO', tk: 'oln4yqj.css' },
+  la: { ietf: 'es-LA', tk: 'oln4yqj.css' },
+  mx: { ietf: 'es-MX', tk: 'oln4yqj.css' },
+  pe: { ietf: 'es-PE', tk: 'oln4yqj.css' },
+  '': { ietf: 'en-US', tk: 'hah7vzn.css' },
+  // EMEA
+  africa: { ietf: 'en', tk: 'pps7abe.css' },
+  be_fr: { ietf: 'fr-BE', tk: 'vrk5vyv.css' },
+  be_en: { ietf: 'en-BE', tk: 'pps7abe.css' },
+  be_nl: { ietf: 'nl-BE', tk: 'cya6bri.css' },
+  cy_en: { ietf: 'en-CY', tk: 'pps7abe.css' },
+  dk: { ietf: 'da-DK', tk: 'aaz7dvd.css' },
+  de: { ietf: 'de-DE', tk: 'vin7zsi.css' },
+  ee: { ietf: 'et-EE', tk: 'aaz7dvd.css' },
+  es: { ietf: 'es-ES', tk: 'oln4yqj.css' },
+  fr: { ietf: 'fr-FR', tk: 'vrk5vyv.css' },
+  gr_en: { ietf: 'en-GR', tk: 'pps7abe.css' },
+  ie: { ietf: 'en-GB', tk: 'pps7abe.css' },
+  il_en: { ietf: 'en-IL', tk: 'pps7abe.css' },
+  it: { ietf: 'it-IT', tk: 'bbf5pok.css' },
+  lv: { ietf: 'lv-LV', tk: 'aaz7dvd.css' },
+  lt: { ietf: 'lt-LT', tk: 'aaz7dvd.css' },
+  lu_de: { ietf: 'de-LU', tk: 'vin7zsi.css' },
+  lu_en: { ietf: 'en-LU', tk: 'pps7abe.css' },
+  lu_fr: { ietf: 'fr-LU', tk: 'vrk5vyv.css' },
+  hu: { ietf: 'hu-HU', tk: 'aaz7dvd.css' },
+  mt: { ietf: 'en-MT', tk: 'pps7abe.css' },
+  mena_en: { ietf: 'en', tk: 'pps7abe.css' },
+  nl: { ietf: 'nl-NL', tk: 'cya6bri.css' },
+  no: { ietf: 'no-NO', tk: 'aaz7dvd.css' },
+  pl: { ietf: 'pl-PL', tk: 'aaz7dvd.css' },
+  pt: { ietf: 'pt-PT', tk: 'inq1xob.css' },
+  ro: { ietf: 'ro-RO', tk: 'aaz7dvd.css' },
+  sa_en: { ietf: 'en', tk: 'pps7abe.css' },
+  ch_de: { ietf: 'de-CH', tk: 'vin7zsi.css' },
+  si: { ietf: 'sl-SI', tk: 'aaz7dvd.css' },
+  sk: { ietf: 'sk-SK', tk: 'aaz7dvd.css' },
+  ch_fr: { ietf: 'fr-CH', tk: 'vrk5vyv.css' },
+  fi: { ietf: 'fi-FI', tk: 'aaz7dvd.css' },
+  se: { ietf: 'sv-SE', tk: 'fpk1pcd.css' },
+  ch_it: { ietf: 'it-CH', tk: 'bbf5pok.css' },
+  tr: { ietf: 'tr-TR', tk: 'aaz7dvd.css' },
+  ae_en: { ietf: 'en', tk: 'pps7abe.css' },
+  uk: { ietf: 'en-GB', tk: 'pps7abe.css' },
+  at: { ietf: 'de-AT', tk: 'vin7zsi.css' },
+  cz: { ietf: 'cs-CZ', tk: 'aaz7dvd.css' },
+  bg: { ietf: 'bg-BG', tk: 'aaz7dvd.css' },
+  ru: { ietf: 'ru-RU', tk: 'aaz7dvd.css' },
+  ua: { ietf: 'uk-UA', tk: 'aaz7dvd.css' },
+  il_he: { ietf: 'he', tk: 'nwq1mna.css', dir: 'rtl' },
+  ae_ar: { ietf: 'ar', tk: 'nwq1mna.css', dir: 'rtl' },
+  mena_ar: { ietf: 'ar', tk: 'dis2dpj.css', dir: 'rtl' },
+  sa_ar: { ietf: 'ar', tk: 'nwq1mna.css', dir: 'rtl' },
+  // Asia Pacific
+  au: { ietf: 'en-AU', tk: 'pps7abe.css' },
+  hk_en: { ietf: 'en-HK', tk: 'pps7abe.css' },
+  in: { ietf: 'en-GB', tk: 'pps7abe.css' },
+  id_id: { ietf: 'id', tk: 'czc0mun.css' },
+  id_en: { ietf: 'en', tk: 'pps7abe.css' },
+  my_ms: { ietf: 'ms', tk: 'sxj4tvo.css' },
+  my_en: { ietf: 'en-GB', tk: 'pps7abe.css' },
+  nz: { ietf: 'en-GB', tk: 'pps7abe.css' },
+  ph_en: { ietf: 'en', tk: 'pps7abe.css' },
+  ph_fil: { ietf: 'fil-PH', tk: 'ict8rmp.css' },
+  sg: { ietf: 'en-SG', tk: 'pps7abe.css' },
+  th_en: { ietf: 'en', tk: 'pps7abe.css' },
+  in_hi: { ietf: 'hi', tk: 'aaa8deh.css' },
+  th_th: { ietf: 'th', tk: 'lqo2bst.css' },
+  cn: { ietf: 'zh-CN', tk: 'puu3xkp' },
+  hk_zh: { ietf: 'zh-HK', tk: 'jay0ecd' },
+  tw: { ietf: 'zh-TW', tk: 'jay0ecd' },
+  jp: { ietf: 'ja-JP', tk: 'dvg6awq' },
+  kr: { ietf: 'ko-KR', tk: 'qjs5sfm' },
+  // Langstore Support.
+  langstore: { ietf: 'en-US', tk: 'hah7vzn.css' },
+  // geo expansion MWPW-125686
+  za: { ietf: 'en-GB', tk: 'pps7abe.css' }, // South Africa (GB English)
+  ng: { ietf: 'en-GB', tk: 'pps7abe.css' }, // Nigeria (GB English)
+  cr: { ietf: 'es-419', tk: 'oln4yqj.css' }, // Costa Rica (Spanish Latin America)
+  ec: { ietf: 'es-419', tk: 'oln4yqj.css' }, // Ecuador (Spanish Latin America)
+  pr: { ietf: 'es-419', tk: 'oln4yqj.css' }, // Puerto Rico (Spanish Latin America)
+  gt: { ietf: 'es-419', tk: 'oln4yqj.css' }, // Guatemala (Spanish Latin America)
+  eg_ar: { ietf: 'ar', tk: 'nwq1mna.css', dir: 'rtl' }, // Egypt (Arabic)
+  kw_ar: { ietf: 'ar', tk: 'nwq1mna.css', dir: 'rtl' }, // Kuwait (Arabic)
+  qa_ar: { ietf: 'ar', tk: 'nwq1mna.css', dir: 'rtl' }, // Qatar (Arabic)
+  eg_en: { ietf: 'en-GB', tk: 'pps7abe.css' }, // Egypt (GB English)
+  kw_en: { ietf: 'en-GB', tk: 'pps7abe.css' }, // Kuwait (GB English)
+  qa_en: { ietf: 'en-GB', tk: 'pps7abe.css' }, // Qatar (GB English)
+  gr_el: { ietf: 'el', tk: 'fnx0rsr.css' }, // Greece (Greek)
+  vn_en: { ietf: 'en-GB', tk: 'hah7vzn.css' },
+  vn_vi: { ietf: 'vi', tk: 'qxw8hzm.css' },
+  cis_ru: { ietf: 'ru', tk: 'qxw8hzm.css' },
+  cis_en: { ietf: 'en', tk: 'pps7abe.css' },
+};
+
+const stageDomainsMap = {
+  'www.stage.adobe.com': {
+    'www.adobe.com': 'origin',
+    'business.adobe.com': 'business.stage.adobe.com',
+    'helpx.adobe.com': 'helpx.stage.adobe.com',
+    'blog.adobe.com': 'blog.stage.adobe.com',
+    'developer.adobe.com': 'developer-stage.adobe.com',
+    'news.adobe.com': 'news.stage.adobe.com',
+    'firefly.adobe.com': 'firefly-stage.corp.adobe.com',
+    'creativecloud.adobe.com': 'stage.creativecloud.adobe.com',
+    'projectneo.adobe.com': 'stg.projectneo.adobe.com',
+  },
+  '--cc--adobecom.hlx.live': {
+    'www.adobe.com': 'origin',
+    'business.adobe.com': 'business.stage.adobe.com',
+    'helpx.adobe.com': 'helpx.stage.adobe.com',
+    'blog.adobe.com': 'blog.stage.adobe.com',
+    'developer.adobe.com': 'developer-stage.adobe.com',
+    'news.adobe.com': 'news.stage.adobe.com',
+    'firefly.adobe.com': 'firefly-stage.corp.adobe.com',
+    'creativecloud.adobe.com': 'stage.creativecloud.adobe.com',
+    'projectneo.adobe.com': 'stg.projectneo.adobe.com',
+  },
+  '--cc--adobecom.hlx.page': {
+    'www.adobe.com': 'origin',
+    'business.adobe.com': 'business.stage.adobe.com',
+    'helpx.adobe.com': 'helpx.stage.adobe.com',
+    'blog.adobe.com': 'blog.stage.adobe.com',
+    'developer.adobe.com': 'developer-stage.adobe.com',
+    'news.adobe.com': 'news.stage.adobe.com',
+    'firefly.adobe.com': 'firefly-stage.corp.adobe.com',
+    'creativecloud.adobe.com': 'stage.creativecloud.adobe.com',
+    'projectneo.adobe.com': 'stg.projectneo.adobe.com',
+  },
+};
 
 /**
  * The decision engine for where to get Milo's libs from.
@@ -195,3 +334,63 @@ export async function acomsisCookieHandler() {
 }
 
 export const decorateArea = getDecorateAreaFn();
+
+const CONFIG = {
+  contentRoot: '/cc-shared',
+  codeRoot: '/creativecloud',
+  imsClientId: 'adobedotcom-cc',
+  locales,
+  geoRouting: 'on',
+  prodDomains: ['www.adobe.com', 'helpx.adobe.com', 'business.adobe.com'],
+  stageDomainsMap,
+  decorateArea,
+  adobeid: {
+    api_parameters: { check_token: { guest_allowed: true } },
+    enableGuestAccounts: true,
+    enableGuestTokenForceRefresh: true,
+  },
+  stage: {
+    pdfViewerClientId: '9f7f19a46bd542e2b8548411e51eb4d4',
+    pdfViewerReportSuite: 'adbadobenonacdcqa',
+    psUrl: 'https://stage.photoshop.adobe.com',
+  },
+  live: {
+    pdfViewerClientId: 'a26c77a2effb4c4aaa71e7c46385e0ed',
+    pdfViewerReportSuite: 'adbadobenonacdcqa',
+  },
+  prod: {
+    pdfViewerClientId: '409019ebd2d546c0be1a0b5a61fe65df',
+    pdfViewerReportSuite: 'adbadobenonacdcprod',
+    psUrl: 'https://photoshop.adobe.com',
+  },
+  jarvis: {
+    id: 'adobedotcom2',
+    version: '1.83',
+    onDemand: false,
+  },
+  htmlExclude: [
+    /www\.adobe\.com\/(\w\w(_\w\w)?\/)?express(\/.*)?/,
+    /www\.adobe\.com\/(\w\w(_\w\w)?\/)?go(\/.*)?/,
+  ],
+};
+
+export const scriptInit = async () => {
+  const isSignedInHomepage = window.location.pathname.includes(CHINA_SIGNED_IN_HOME_PATH);
+  const { loadArea, setConfig, loadLana } = await import(`${miloLibs}/utils/utils.js`);
+  setConfig({ ...CONFIG, miloLibs });
+  if (isSignedInHomepage) acomsisCookieHandler();
+  decorateArea();
+  (function loadStyles() {
+    const paths = [`${miloLibs}/styles/styles.css`];
+    paths.forEach((path) => {
+      const link = document.createElement('link');
+      link.setAttribute('rel', 'stylesheet');
+      link.setAttribute('href', path);
+      document.head.appendChild(link);
+    });
+  }());
+  (async function loadPage() {
+    loadLana({ clientId: 'cc' });
+    await loadArea();
+  }());
+};
