--- creativecloud/styles/styles.css
@@ -1,16 +1,0 @@
- .reading-width {
-  max-width: 800px;
-  margin: auto;
-  padding: 20px;
-}
-
- .reading-width-footer-tag {
-  max-width: 800px;
-  margin: -20px auto;
-  padding: 0 20px 20px;
-}
-
- .reading-width-footer-tag > .content {
-  max-width: var(--grid-container-width);
-  margin: 0 auto;
-}
